dict to list

+++ inference.py
@@ -231,7 +231,7 @@ def run_one_inference(img, model, args, name, logger, show=True, show_depth=True
     floor_pts = []
     for pt in json_data['layoutPoints']['points']:
         floor_pts.append(xyz2pixel(np.array(pt['xyz']),w=256,h=128))
-    floor_pts = np.round(np.array(floor_pts))
+    floor_pts = np.round(np.array(floor_pts)).astype(np.int64)
     floor_pts = floor_pts[::-1]
     min_value = np.min(floor_pts[:,0])
     loop_cnt = 0
